refactor: 전체 게시글 섹션 컴포넌트로 분리
