+++ components/AllArticleSection.module.css
@@ -1,0 +1,88 @@
+.articleContainer{
+  margin: 24px auto 16px;
+}
+
+.articleHeader {
+  width: 1200px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 24px;
+}
+
+.articleTitle {
+  font-size: 20px;
+  font-weight: 700;
+  line-height: 24px;
+  color: #111827;
+}
+
+.articleContentBottom {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.likeCountWrapper {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.writeButton {
+  width: 88px;
+  height: 42px;
+  border-radius: 8px;
+  border: none;
+  background-color: #3692ff;
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 19px;
+  color: #fff;
+}
+
+.searchInputWrapper{
+  position: relative;
+}
+
+.searchInput {
+  width: 1054px;
+  height: 42px;
+  border-radius: 12px;
+  border: none;
+  background-color: #f3f4f6;
+  padding: 9px 44px;
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 24px;
+}
+
+.sortButton{
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  width: 130px;
+  height: 42px;  
+  border-radius: 12px;
+  border: 1px solid #e5e7eb;
+  background-color: #fff;
+}
+
+.allArticleWrapper{
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  width: 1200px;
+  height: 136px;
+  border-bottom: 1px solid #e5e7eb;
+  border-radius: 1px;
+  background-color: #fff;
+  padding: 24px 0;
+}
+
+.allArticleBottomLeft{
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
